fix: rename Helmet to be a client component
